Create socket validations/Socket rearranged/Upload images/Improved code

+++ backend/src/socket/admin-sockets.ts
@@ -1,13 +1,16 @@
-import type { IGroup, Members } from '../types/global.js';
+import type { Members } from '../types/global.js';
 import type { AdminSockets } from '../types/sockets.js';
+import fs from 'fs-extra';
+import { resolve } from 'path';
 import { User, Group, Chat } from '../models/index.js';
+import { TypeContact } from '../types/enums.js';
 
 export const adminSockets: AdminSockets = (socket, [userID, contactID], groupRooms) => {
 	socket.on('emitAddMod', async (mods: Members[]) => {
 		const userIDs = mods.map(mods => mods.id);
 
 		await Group.updateOne(
-			{ _id: contactID },
+			{ _id: contactID, admin: userID },
 			{
 				$push: { mods },
 				$pull: { members: { id: { $in: userIDs } } }
@@ -21,7 +24,7 @@ export const adminSockets: AdminSockets = (socket, [userID, contactID], groupRoo
 		const userIDs = members.map(member => member.id);
 
 		await Group.updateOne(
-			{ _id: contactID },
+			{ _id: contactID, admin: userID },
 			{
 				$push: { members },
 				$pull: { mods: { id: { $in: userIDs } } }
@@ -31,27 +34,41 @@ export const adminSockets: AdminSockets = (socket, [userID, contactID], groupRoo
 		socket.to(contactID).emit('modChange', userID);
 	});
 
+	socket.on('emitChangeAvatar', async (filename: string) => {
+		socket.emit('changeAvatar', contactID, TypeContact.GROUP, filename);
+		socket.to(groupRooms).emit('changeAvatar', contactID, TypeContact.GROUP, filename);
+	});
+
+	socket.on('emitChangeDescription', async (description: string) => {
+		await Group.updateOne({ _id: contactID, admin: userID }, { description });
+	});
+
 	socket.on('emitChangeState', async (state: string) => {
-		await Group.updateOne({ _id: contactID }, { state });
+		await Group.updateOne({ _id: contactID, admin: userID }, { state });
 	});
 
 	socket.on('emitDestroyGroup', async () => {
-		const { admin, mods, members } = await Group
-			.findOneAndDelete({ _id: contactID })
-			.select('admin mods members') as IGroup;
-			
-		const contactsIDs = [...mods, ...members].map(user => user.id);
+		const group = await Group
+			.findOneAndDelete({ _id: contactID, admin: userID })
+			.lean({ virtuals: true });
 
-		[admin, ...contactsIDs].forEach(async id => {
-			await User.updateOne({ _id: id }, { $pull: { groupRooms: contactID } });
-		});
+		if (group !== null) {
+			[group.admin, ...group.modIDs, ...group.memberIDs].forEach(async id => {
+				await User.updateOne({ _id: id }, { $pull: { groupRooms: contactID } });
+			});
 
-		await Chat.deleteMany({ to: contactID });
+			await Chat.deleteMany({ to: contactID });
 
-		groupRooms = groupRooms.filter(id => id !== contactID);
+			if (group.avatar !== 'avatar.jpeg') {
+				const path = resolve(`uploads/group-avatar/${group.avatar}`);
+				await fs.unlink(path);
+			}
+
+			groupRooms = groupRooms.filter(id => id !== contactID);
 		
-		socket.to(contactID).emit('leaveGroup', contactID);
-		socket.leave(contactID);
+			socket.to(contactID).emit('leaveGroup', contactID);
+			socket.leave(contactID);
+		}
 	});
 
 	return groupRooms;
